requirements.txt
PyQt5
flet>=0.10.0
requests>=2.28.1
cryptography>=40.0.0
pycryptodome>=3.17.0
pillow>=9.4.0
psutil>=5.9.4
pywin32>=306; platform_system == "Windows"
pyinstaller>=5.11.0; platform_system == "Windows"
opencv-python>=4.8.0; platform_system == "Windows"
pyautogui>=0.9.54
# Зависимости для улучшенного обфускатора
ast-decompiler>=0.4.0
